refactor(cli): split upgrade into finalization step and perform with new version
