+++ packages/@expo/cli/src/upgrade/__tests__/finalizeAsync.test.ts
@@ -1,4 +1,4 @@
-import { getPackagesToModify } from '../upgradeAsync';
+import { getPackagesToModify } from '../finalizeAsync';
 
 describe(getPackagesToModify, () => {
   it(`should remove deprecated packages`, () => {
